RBM : Shuffle database
This commit answer to a task of #1
RBM is trained one by one using gesture after another. This principle
could create a bias cos the last gesture could impact quiet more the
weights of the RBM than the other gestures. In order to delete this
prospective bias, we shuffle the database. Now, all gesture are learned
in the same time and have the same influence on the RBM’s weights.